// Eurocraft/src/app/products/models/productcategory.ts
export class ProductCategory {
    ProductCategoryId: number;
    ProductCategoryName: string;
    ProductMaterialId: number | null;
    ProductMaterialName: string;
    ProductModelId: number | null;
    ProductModelName: string;
    ProductTypeId: number;
    ProductTypeName: string;
    ProductUseId: number | null;
    ProductUseName: string;
}